feat(ancillary): replace accordion with AncillarySelection component for service selection
